[V6] Update coron-4.1 tools to V6

+++ target_files_from_device.sh
@@ -11,6 +11,8 @@ PRJ_ROOT=`pwd`
 TOOL_DIR=$PORT_ROOT/tools
 TARGET_FILES_TEMPLATE_DIR=$TOOL_DIR/target_files_template
 OTA_FROM_TARGET_FILES=$TOOL_DIR/releasetools/ota_from_target_files
+SYSTEM_INFO_PROCESS=$TOOL_DIR/releasetools/systeminfoprocess.py
+RECOVERY_LINK=$TOOL_DIR/releasetools/recoverylink.py
 
 OUT_DIR=$PRJ_ROOT/out
 OEM_TARGET_DIR=$OUT_DIR/oem_target_files
@@ -20,12 +22,35 @@ RECOVERY_ETC_DIR=$OEM_TARGET_DIR/RECOVERY/RAMDISK/etc
 OEM_TARGET_ZIP=$OUT_DIR/oem_target_files.zip
 OUTPUT_OTA_PACKAGE=$OUT_DIR/oem_ota_rom.zip
 
+ROOT_STATE="system_root"
+
+# check system root state
+function check_root_state {
+	echo ">>> check root state of phone ..."
+	wait_for_device_online
+	SECURE_PROP=$(adb shell cat /default.prop | grep -o "ro.secure=\w")
+	DEBUG_PROP=$(adb shell cat /default.prop | grep -o "ro.debuggable=\w")
+	if [ "$SECURE_PROP" = "ro.secure=0" -o "$DEBUG_PROP" = "ro.debuggable=1" ];then
+		ROOT_STATE="kernel_root"
+		echo ">>> system root state: kernel root"
+		adb root
+		wait_for_device_online
+	else
+		echo "exit" > exit_command
+		wait_for_device_online
+		adb push exit_command /data/local/tmp
+		rm -f exit_command
+		if echo "su < /data/local/tmp/exit_command; exit" | adb shell | grep "not found" > /dev/null;then
+			echo ">>> ERROR: Not a root phone, please root this device firstly"
+			exit 1;
+		fi
+		ROOT_STATE="system_root"
+		echo ">>> system root state: system root"
+	fi
+}
+
 # check for files preparing
 function check_for_env_prepare {
-	if [ ! -f $PRJ_ROOT/boot.img ];then
-		echo ">>> can not find $PRJ_ROOT/boot.img"
-		exit 1
-	fi
 	if [ ! -f $PRJ_ROOT/recovery.fstab ];then
 		echo ">>> can not find $PRJ_ROOT/recovery.fstab"
 		exit 1
@@ -64,30 +89,20 @@ function wait_for_device_online {
 
 # get device system files info
 function get_device_system_info {
-	adb push $TOOL_DIR/releasetools/getfilesysteminfo /data/local/tmp
-	adb shell chmod 0777 /data/local/tmp/getfilesysteminfo
+	echo ">>> get device system info ..."
+	adb push $TOOL_DIR/releasetools/getfilesysteminfo.sh /data/local/tmp
 
-	if adb root | grep "cannot run" > /dev/null;then
-		echo ">>> system is not kernel root, check weather it can run in su mode"
-		echo "exit" > exit_command
-		wait_for_device_online
-		adb push exit_command /data/local/tmp
-		rm -f exit_command
-		if echo "su < /data/local/tmp/exit_command; exit" | adb shell | grep "not found" > /dev/null;then
-			echo ">>> ERROR: can not find su command, please root this device firstly"
-			exit 1;
-		else
-			wait_for_device_online
-			adb push $TOOL_DIR/releasetools/getsysteminfocommand /data/local/tmp
-			echo "su < /data/local/tmp/getsysteminfocommand; exit" | adb shell
-			adb pull /data/local/tmp/system.info $META_DIR/
-			adb pull /data/local/tmp/link.info $META_DIR/
-		fi
+	wait_for_device_online
+	if [ "$ROOT_STATE" = "system_root" ];then
+		adb push $TOOL_DIR/releasetools/getsysteminfocommand /data/local/tmp
+		echo "su < /data/local/tmp/getsysteminfocommand; exit" | adb shell
 	else
-		wait_for_device_online
-		adb shell /data/local/tmp/getfilesysteminfo --info /system > $META_DIR/system.info
-		adb shell /data/local/tmp/getfilesysteminfo --link /system > $META_DIR/link.info
+		adb shell chmod 0777 /data/local/tmp/getfilesysteminfo.sh
+		adb shell /data/local/tmp/getfilesysteminfo.sh
 	fi
+	adb pull /data/local/tmp/file.info $META_DIR/
+	$SYSTEM_INFO_PROCESS $META_DIR/file.info $META_DIR/system.info $META_DIR/link.info
+	rm -f $META_DIR/file.info
 
 	if [ -f $META_DIR/system.info -a -f $META_DIR/link.info ];then
 		echo ">>> get device system files info done"
@@ -122,21 +137,86 @@ function build_filesystem_config {
 # recover the device files' symlink information
 function recover_symlink {
     echo ">>> Run recoverylink.py to recover symlink"
-    cat $META_DIR/link.info | sed -e '/\<su\>/d;/\<invoke-as\>/d' | sort > $SYSTEM_DIR/linkinfo.txt
-    python $TOOL_DIR/releasetools/recoverylink.py $OEM_TARGET_DIR
-	mv $SYSTEM_DIR/linkinfo.txt $META_DIR
+    cat $META_DIR/link.info | sed -e '/\<su\>/d;/\<invoke-as\>/d' | sort > $META_DIR/linkinfo.txt
+    $RECOVERY_LINK $META_DIR/linkinfo.txt $OEM_TARGET_DIR
 	if [ ! -f $META_DIR/linkinfo.txt ];then
 		echo ">>> Failed to create linkinfo.txt"
 		exit 1
 	fi
 }
 
+function turnModToNum {
+    mod=$1
+    ((num=2#$(echo $mod | sed 's/[^\-]/1/g; s/\-/0/g')))
+    echo $num
+}
+
+function addReadMod {
+    sFile=$1
+
+    sFileMod=$(adb shell ls -l $sFile | awk '{print $1}')
+    ownMod=${sFileMod:1:3}
+    grpMod=${sFileMod:4:3}
+    otherMod="r"${sFileMod:8:2}
+
+    newMod=$(turnModToNum $ownMod)$(turnModToNum $grpMod)$(turnModToNum $otherMod)
+
+    if [ "x$newMod" != "x" ];then
+        if [ x"$ROOT_STATE" = x"system_root" ]; then
+            adb shell "su -c \"mount -o remount,rw /system; chmod $newMod $sFile; exit;\""
+        else
+            echo "mount -o remount,rw /system; chmod $newMod $sFile;exit" | adb shell
+        fi
+    fi
+}
+
+function pullFailedFailes {
+    failedLog=$1
+    times=$2
+    tmpLog=$(mktemp -t "tmp.pull.XXXXX")
+
+    grep "^failed to copy" $failedLog | while read LINE
+    do
+        sFile=$(echo $LINE | awk -F \' '{print $2}')
+        outFile=$(echo $LINE | awk -F \' '{print $4}')
+
+        addReadMod $sFile
+        adb pull $sFile $outFile 2>&1 | grep "^failed to copy" | tee $tmpLog
+    done
+
+    if [ $(test -s $tmpLog) ]; then
+         times=$(expr $times - 1)
+         if [ $times -gt 0 ]; then
+             pullFailedFailes $tmpLog
+         else
+             cat $tmpLog > $OUT_DIR/system-pull-failed.log
+         fi
+    fi
+    rm $tmpLog
+}
+
+function deal_with_system_pull_log {
+	pullLog=$1
+	pullFailedFailes $pullLog 20
+	if [ -s $OUT_DIR/system-pull-failed.log ];then
+		echo "-------------------------------------------------------" > $OUT_DIR/build-info-to-user.txt
+		echo "Some files those pull failed you must deal with manually:" >> $OUT_DIR/build-info-to-user.txt
+		cat $OUT_DIR/system-pull-failed.log | sed -e "s/.*out\/oem_target_files\/SYSTEM\/\(.*\)'.*/\1/" >> $OUT_DIR/build-info-to-user.txt
+		echo "" >> $OUT_DIR/build-info-to-user.txt
+		echo "---------" >> $OUT_DIR/build-info-to-user.txt
+		echo "pull log:" >> $OUT_DIR/build-info-to-user.txt
+		cat $OUT_DIR/system-pull-failed.log >> $OUT_DIR/build-info-to-user.txt
+		echo "-------------------------------------------------------" >> $OUT_DIR/build-info-to-user.txt
+	fi
+}
+
 # build the SYSTEM dir under target_files
 function build_SYSTEM {
     echo ">>> Extract the whole /system from device"
-    adb pull /system $SYSTEM_DIR
+    adb pull /system $SYSTEM_DIR 2>&1 | tee $OUT_DIR/system-pull.log
     find $SYSTEM_DIR -name su | xargs rm -f
     find $SYSTEM_DIR -name invoke-as | xargs rm -f
+	deal_with_system_pull_log $OUT_DIR/system-pull.log
 
     build_filesystem_config
     recover_symlink
@@ -145,6 +225,14 @@ function build_SYSTEM {
 # build apkcerts.txt from packages.xml
 function build_apkcerts {
     echo ">>> Build apkcerts.txt"
+	if [ x"$ROOT_STATE" = x"system_root" ];then
+		echo "chmod 666 /data/system/packages.xml" > chmodcommand
+		adb push chmodcommand /data/local/tmp/chmodcommand
+		rm chmodcommand
+		echo "su < /data/local/tmp/chmodcommand; exit" | adb shell
+	else
+		adb shell chmod 666 /data/system/packages.xml
+	fi
     adb pull /data/system/packages.xml $OEM_TARGET_DIR
     python $TOOL_DIR/apkcerts.py $OEM_TARGET_DIR/packages.xml $META_DIR/apkcerts.txt
     for file in `ls $SYSTEM_DIR/framework/*.apk`
@@ -163,12 +251,16 @@ function build_apkcerts {
 
 # prepare boot.img recovery.fstab for target
 function prepare_boot_recovery {
-	mkdir -p $OEM_TARGET_DIR/BOOTABLE_IMAGES
-	cp -f $PRJ_ROOT/boot.img $OEM_TARGET_DIR/BOOTABLE_IMAGES/boot.img
+	if [ -f $PRJ_ROOT/boot.img ];then
+		mkdir -p $OEM_TARGET_DIR/BOOTABLE_IMAGES
+		cp -f $PRJ_ROOT/boot.img $OEM_TARGET_DIR/BOOTABLE_IMAGES/boot.img
+		echo ">>> Copy boot.img to $OEM_TARGET_DIR/BOOTABLE_IMAGES/boot.img"
+	fi
 	if [ ! -d $RECOVERY_ETC_DIR ];then
 		mkdir -p $RECOVERY_ETC_DIR
 	fi
 	cp -f $PRJ_ROOT/recovery.fstab $RECOVERY_ETC_DIR/recovery.fstab
+	echo ">>> Copy recovery.fstab to $RECOVERY_ETC_DIR/recovery.fstab"
 }
 
 # compress the target_files dir into a zip file
@@ -186,6 +278,7 @@ function zip_target_files {
 
 # pull files from phone and build a target file
 function target-from-phone {
+	check_root_state
 	check_for_env_prepare
 	copy_target_files_template
 	wait_for_device_online
